Site Editor: Add the Post Author component to the Page sidebar (#56368)
