--- packages/edit-post/src/components/sidebar/post-author/style.scss
@@ -1,5 +1,0 @@
-.edit-post-post-author {
-	display: flex;
-	flex-direction: column;
-	align-items: stretch;
-}
