+++ packages/edit-post/src/style.scss
@@ -10,7 +10,6 @@
 @import "./components/secondary-sidebar/style.scss";
 @import "./components/sidebar/style.scss";
 @import "./components/sidebar/last-revision/style.scss";
-@import "./components/sidebar/post-author/style.scss";
 @import "./components/sidebar/post-format/style.scss";
 @import "./components/sidebar/post-slug/style.scss";
 @import "./components/sidebar/post-template/style.scss";
