+++ packages/editor/src/components/post-author/style.scss
@@ -1,0 +1,7 @@
+.editor-post-author__panel {
+	padding-top: $grid-unit-10;
+}
+
+.editor-post-author__panel .editor-post-panel__row-control > div {
+	width: 100%;
+}
